+++ modules/datadog/widget.go
@@ -2,7 +2,6 @@ package datadog
 
 import (
 	"fmt"
-	"strconv"
 
 	"github.com/rivo/tview"
 	"github.com/wtfutil/wtf/wtf"
@@ -24,32 +23,27 @@ const HelpText = `
 
 type Widget struct {
 	wtf.HelpfulWidget
-	wtf.TextWidget
 	wtf.KeyboardWidget
+	wtf.ScrollableWidget
 
-	app      *tview.Application
-	selected int
 	monitors []datadog.Monitor
 	settings *Settings
 }
 
 func NewWidget(app *tview.Application, pages *tview.Pages, settings *Settings) *Widget {
 	widget := Widget{
-		HelpfulWidget:  wtf.NewHelpfulWidget(app, pages, HelpText),
-		KeyboardWidget: wtf.NewKeyboardWidget(),
-		TextWidget:     wtf.NewTextWidget(app, settings.common, true),
+		HelpfulWidget:    wtf.NewHelpfulWidget(app, pages, HelpText),
+		KeyboardWidget:   wtf.NewKeyboardWidget(),
+		ScrollableWidget: wtf.NewScrollableWidget(app, settings.common, true),
 
 		app:      app,
 		settings: settings,
 	}
 
+	widget.SetRenderFunction(widget.Render)
 	widget.initializeKeyboardControls()
 	widget.View.SetInputCapture(widget.InputCapture)
 
-	widget.unselect()
-
-	widget.View.SetScrollable(true)
-	widget.View.SetRegions(true)
 	widget.HelpfulWidget.SetView(widget.View)
 
 	return &widget
@@ -60,14 +54,10 @@ func NewWidget(app *tview.Application, pages *tview.Pages, settings *Settings) *
 func (widget *Widget) Refresh() {
 	monitors, monitorErr := widget.Monitors()
 
-	var content string
 	if monitorErr != nil {
-		widget.View.SetWrap(true)
-		content = monitorErr.Error()
-		widget.app.QueueUpdateDraw(func() {
-			widget.View.SetTitle(widget.ContextualTitle(widget.CommonSettings.Title))
-			widget.View.SetText(content)
-		})
+		widget.monitors = nil
+		widget.SetItemCount(0)
+		widget.Redraw(widget.CommonSettings.Title, monitorErr.Error(), true)
 		return
 	}
 	triggeredMonitors := []datadog.Monitor{}
@@ -80,19 +70,14 @@ func (widget *Widget) Refresh() {
 		}
 	}
 	widget.monitors = triggeredMonitors
+	widget.SetItemCount(len(widget.monitors))
 
-	widget.display()
+	widget.Render()
 }
 
-func (widget *Widget) display() {
+func (widget *Widget) Render() {
 	content := widget.contentFrom(widget.monitors)
-	widget.app.QueueUpdateDraw(func() {
-		widget.View.SetWrap(false)
-		widget.View.Clear()
-		widget.View.SetTitle(widget.ContextualTitle(widget.CommonSettings.Title))
-		widget.View.SetText(content)
-		widget.View.Highlight(strconv.Itoa(widget.selected)).ScrollToHighlight()
-	})
+	widget.Redraw(widget.CommonSettings.Title, content, false)
 }
 
 /* -------------------- Unexported Functions -------------------- */
@@ -108,9 +93,9 @@ func (widget *Widget) contentFrom(triggeredMonitors []datadog.Monitor) string {
 		for idx, triggeredMonitor := range triggeredMonitors {
 			str = str + fmt.Sprintf(`["%d"][%s][red] %s[%s][""]`,
 				idx,
-				widget.rowColor(idx),
+				widget.RowColor(idx),
 				*triggeredMonitor.Name,
-				widget.rowColor(idx),
+				widget.RowColor(idx),
 			) + "\n"
 		}
 	} else {
@@ -123,40 +108,11 @@ func (widget *Widget) contentFrom(triggeredMonitors []datadog.Monitor) string {
 	return str
 }
 
-func (widget *Widget) unselect() {
-	widget.selected = -1
-	widget.display()
-}
-
-func (widget *Widget) next() {
-	widget.selected++
-	if widget.monitors != nil && widget.selected >= len(widget.monitors) {
-		widget.selected = 0
-	}
-	widget.display()
-}
-
-func (widget *Widget) prev() {
-	widget.selected--
-	if widget.selected < 0 && widget.monitors != nil {
-		widget.selected = len(widget.monitors) - 1
-	}
-	widget.display()
-}
-
 func (widget *Widget) openItem() {
 
-	sel := widget.selected
+	sel := widget.GetSelected()
 	if sel >= 0 && widget.monitors != nil && sel < len(widget.monitors) {
-		item := &widget.monitors[widget.selected]
+		item := &widget.monitors[sel]
 		wtf.OpenFile(fmt.Sprintf("https://app.datadoghq.com/monitors/%d?q=*", *item.Id))
 	}
-}
-
-func (widget *Widget) rowColor(idx int) string {
-	if widget.View.HasFocus() && (idx == widget.selected) {
-		widget.settings.common.DefaultFocussedRowColor()
-	}
-
-	return widget.settings.common.RowColor(idx)
 }