+++ modules/gitter/widget.go
@@ -2,7 +2,6 @@ package gitter
 
 import (
 	"fmt"
-	"strconv"
 
 	"github.com/rivo/tview"
 	"github.com/wtfutil/wtf/wtf"
@@ -24,33 +23,25 @@ const HelpText = `
 type Widget struct {
 	wtf.HelpfulWidget
 	wtf.KeyboardWidget
-	wtf.TextWidget
-
-	app *tview.Application
+	wtf.ScrollableWidget
 
 	messages []Message
-	selected int
 	settings *Settings
 }
 
 // NewWidget creates a new instance of a widget
 func NewWidget(app *tview.Application, pages *tview.Pages, settings *Settings) *Widget {
 	widget := Widget{
-		HelpfulWidget:  wtf.NewHelpfulWidget(app, pages, HelpText),
-		KeyboardWidget: wtf.NewKeyboardWidget(),
-		TextWidget:     wtf.NewTextWidget(app, settings.common, true),
+		HelpfulWidget:    wtf.NewHelpfulWidget(app, pages, HelpText),
+		KeyboardWidget:   wtf.NewKeyboardWidget(),
+		ScrollableWidget: wtf.NewScrollableWidget(app, settings.common, true),
 
-		app:      app,
 		settings: settings,
 	}
 
+	widget.SetRenderFunction(widget.Refresh)
 	widget.initializeKeyboardControls()
 	widget.View.SetInputCapture(widget.InputCapture)
-
-	widget.unselect()
-
-	widget.View.SetScrollable(true)
-	widget.View.SetRegions(true)
 
 	widget.HelpfulWidget.SetView(widget.View)
 
@@ -66,25 +57,23 @@ func (widget *Widget) Refresh() {
 
 	room, err := GetRoom(widget.settings.roomURI, widget.settings.apiToken)
 	if err != nil {
-		widget.View.SetWrap(true)
-		widget.View.SetTitle(widget.CommonSettings.Title)
-		widget.View.SetText(err.Error())
+		widget.Redraw(widget.CommonSettings.Title, err.Error(), true)
 		return
 	}
 
 	if room == nil {
+		widget.Redraw(widget.CommonSettings.Title, "No room", true)
 		return
 	}
 
 	messages, err := GetMessages(room.ID, widget.settings.numberOfMessages, widget.settings.apiToken)
 
 	if err != nil {
-		widget.View.SetWrap(true)
-
 		widget.Redraw(widget.CommonSettings.Title, err.Error(), true)
 		return
 	}
 	widget.messages = messages
+	widget.SetItemCount(len(messages))
 
 	widget.display()
 }
@@ -99,10 +88,6 @@ func (widget *Widget) display() {
 	title := fmt.Sprintf("%s - %s", widget.CommonSettings.Title, widget.settings.roomURI)
 
 	widget.Redraw(title, widget.contentFrom(widget.messages), true)
-	widget.app.QueueUpdateDraw(func() {
-		widget.View.Highlight(strconv.Itoa(widget.selected)).ScrollToHighlight()
-		widget.View.ScrollToEnd()
-	})
 }
 
 func (widget *Widget) contentFrom(messages []Message) string {
@@ -111,10 +96,10 @@ func (widget *Widget) contentFrom(messages []Message) string {
 		str = str + fmt.Sprintf(
 			`["%d"][%s] [blue]%s [lightslategray]%s: [%s]%s [aqua]%s[""]`,
 			idx,
-			widget.rowColor(idx),
+			widget.RowColor(idx),
 			message.From.DisplayName,
 			message.From.Username,
-			widget.rowColor(idx),
+			widget.RowColor(idx),
 			message.Text,
 			message.Sent.Format("Jan 02, 15:04 MST"),
 		)
@@ -125,41 +110,10 @@ func (widget *Widget) contentFrom(messages []Message) string {
 	return str
 }
 
-func (widget *Widget) rowColor(idx int) string {
-	if widget.View.HasFocus() && (idx == widget.selected) {
-		return widget.settings.common.DefaultFocussedRowColor()
-	}
-
-	return widget.settings.common.RowColor(idx)
-}
-
-func (widget *Widget) next() {
-	widget.selected++
-	if widget.messages != nil && widget.selected >= len(widget.messages) {
-		widget.selected = 0
-	}
-
-	widget.display()
-}
-
-func (widget *Widget) prev() {
-	widget.selected--
-	if widget.selected < 0 && widget.messages != nil {
-		widget.selected = len(widget.messages) - 1
-	}
-
-	widget.display()
-}
-
 func (widget *Widget) openMessage() {
-	sel := widget.selected
+	sel := widget.GetSelected()
 	if sel >= 0 && widget.messages != nil && sel < len(widget.messages) {
-		message := &widget.messages[widget.selected]
+		message := &widget.messages[sel]
 		wtf.OpenFile(message.Text)
 	}
-}
-
-func (widget *Widget) unselect() {
-	widget.selected = -1
-	widget.display()
 }