+++ modules/hackernews/keyboard.go
@@ -1,18 +1,17 @@
-
 package hackernews
 
 import "github.com/gdamore/tcell"
 
 func (widget *Widget) initializeKeyboardControls() {
 	widget.SetKeyboardChar("/", widget.ShowHelp)
-	widget.SetKeyboardChar("j", widget.next)
-	widget.SetKeyboardChar("k", widget.prev)
+	widget.SetKeyboardChar("j", widget.Next)
+	widget.SetKeyboardChar("k", widget.Prev)
 	widget.SetKeyboardChar("o", widget.openStory)
 	widget.SetKeyboardChar("r", widget.Refresh)
 	widget.SetKeyboardChar("c", widget.openComments)
 
-	widget.SetKeyboardKey(tcell.KeyDown, widget.next)
+	widget.SetKeyboardKey(tcell.KeyDown, widget.Next)
 	widget.SetKeyboardKey(tcell.KeyEnter, widget.openStory)
-	widget.SetKeyboardKey(tcell.KeyEsc, widget.unselect)
-	widget.SetKeyboardKey(tcell.KeyUp, widget.prev)
+	widget.SetKeyboardKey(tcell.KeyEsc, widget.Unselect)
+	widget.SetKeyboardKey(tcell.KeyUp, widget.Prev)
 }